move unit tests to tests/tracer
