+++ tests/tracer/test_ext.py
