+++ tox.ini
@@ -150,8 +150,6 @@ envlist =
     {py27}-opentracer_gevent-gevent{10}
     {py27,py35,py36}-opentracer_gevent-gevent{11,12}
     py{37,38}-opentracer_gevent-gevent{13,14}
-# Unit tests: pytest based test suite that do not require any additional dependency
-    unit_tests-{py27,py35,py36,py37,py38}
     benchmarks-{py27,py35,py36,py37,py38}
 
 isolated_build = true
@@ -534,8 +532,6 @@ commands =
     ddtracerun: pytest {posargs} tests/commands/test_runner.py
     test_utils: pytest {posargs} tests/contrib/test_utils.py
     test_logging: pytest {posargs} tests/contrib/logging/
-# Unit tests: pytest based test suite that do not require any additional dependency.
-    unit_tests: pytest {posargs} tests/unit
     benchmarks: pytest {posargs} tests/benchmark.py
 
 [testenv:wait]
